- Optimize Tabview and BottomNavigation package
