+++ app/src/main/java/com/piashcse/hilt_mvvm_compose_movie/ui/screens/mainscreen/MainScreen.kt
@@ -1,10 +1,8 @@
 package com.piashcse.hilt_mvvm_compose_movie.ui.screens.mainscreen
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.PagerState
@@ -19,31 +17,22 @@ import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Snackbar
-import androidx.compose.material3.Tab
-import androidx.compose.material3.TabRow
-import androidx.compose.material3.TabRowDefaults
-import androidx.compose.material3.TabRowDefaults.tabIndicatorOffset
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberTopAppBarState
-import androidx.compose.runtime.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.piashcse.hilt_mvvm_compose_movie.R
@@ -56,6 +45,9 @@ import com.piashcse.hilt_mvvm_compose_movie.navigation.navigationTitle
 import com.piashcse.hilt_mvvm_compose_movie.ui.component.CircularIndeterminateProgressBar
 import com.piashcse.hilt_mvvm_compose_movie.ui.component.SearchBar
 import com.piashcse.hilt_mvvm_compose_movie.ui.component.SearchUI
+import com.piashcse.hilt_mvvm_compose_movie.ui.screens.mainscreen.botton_navigation.BottomNavigationUI
+import com.piashcse.hilt_mvvm_compose_movie.ui.screens.mainscreen.tav_view.FavoriteTabView
+import com.piashcse.hilt_mvvm_compose_movie.ui.screens.mainscreen.tav_view.MovieTvSeriesTabView
 import com.piashcse.hilt_mvvm_compose_movie.ui.theme.FloatingActionBackground
 import com.piashcse.hilt_mvvm_compose_movie.ui.theme.cornerRadius
 import com.piashcse.hilt_mvvm_compose_movie.utils.ACTIVE_MOVIE_TAB
@@ -65,9 +57,7 @@ import com.piashcse.hilt_mvvm_compose_movie.utils.network.DataState
 import com.piashcse.hilt_mvvm_compose_movie.utils.networkconnection.ConnectionState
 import com.piashcse.hilt_mvvm_compose_movie.utils.networkconnection.connectivityState
 import com.piashcse.hilt_mvvm_compose_movie.utils.pagingLoadingState
-import com.piashcse.hilt_mvvm_compose_movie.utils.singleTopNavigator
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalCoroutinesApi::class)
 @Composable
@@ -215,36 +205,6 @@ fun MainScreen() {
     }
 }
 
-
-@Composable
-fun BottomNavigationUI(navController: NavController, pagerState: PagerState) {
-    NavigationBar {
-        val items = if (pagerState.currentPage == 0) {
-            listOf(
-                Screen.NowPlayingNav,
-                Screen.PopularNav,
-                Screen.TopRatedNav,
-                Screen.UpcomingNav,
-            )
-        } else {
-            listOf(
-                Screen.AiringTodayTvSeriesNav,
-                Screen.OnTheAirTvSeriesNav,
-                Screen.PopularTvSeriesNav,
-                Screen.TopRatedTvSeriesNav,
-            )
-        }
-        items.forEachIndexed { index, item ->
-            NavigationBarItem(icon = item.navIcon,
-                label = { Text(text = stringResource(id = item.title)) },
-                selected = currentRoute(navController) == item.route,
-                onClick = {
-                    navController.singleTopNavigator(item.route)
-                })
-        }
-    }
-}
-
 @Composable
 fun MainView(
     navigator: NavHostController,
@@ -262,79 +222,6 @@ fun MainView(
             state = pagerState, modifier = Modifier.fillMaxSize()
         ) { page ->
             Navigation(navigator, genres)
-        }
-    }
-}
-
-@Composable
-fun MovieTvSeriesTabView(
-    navigator: NavHostController,
-    pagerState: PagerState,
-) {
-    val coroutineScope = rememberCoroutineScope()
-    val tabs = listOf(stringResource(R.string.movie), stringResource(R.string.tv_series))
-    TabRow(modifier = Modifier.background(Color.White),
-        selectedTabIndex = pagerState.currentPage,
-        indicator = { tabPositions ->
-            TabRowDefaults.PrimaryIndicator(
-                Modifier.tabIndicatorOffset(tabPositions[pagerState.currentPage]),
-                color = MaterialTheme.colorScheme.primary
-            )
-        }) {
-        tabs.forEachIndexed { index, title ->
-            Tab(selected = pagerState.currentPage == index, onClick = {
-                if (index == ACTIVE_MOVIE_TAB) {
-                    navigator.singleTopNavigator(Screen.NowPlaying.route)
-
-                } else if (index == ACTIVE_TV_SERIES_TAB) {
-                    navigator.singleTopNavigator(Screen.AiringTodayTvSeries.route)
-                }
-                coroutineScope.launch {
-                    pagerState.animateScrollToPage(index)
-                }
-            }, text = {
-                Text(
-                    title,
-                    color = if (pagerState.currentPage == index) MaterialTheme.colorScheme.primary else Color.Gray
-                )
-            })
-        }
-    }
-}
-
-@Composable
-fun FavoriteTabView(navigator: NavHostController) {
-    var selectedTabIndex by remember { mutableIntStateOf(0) }
-    val tabs = listOf(stringResource(R.string.movie), stringResource(R.string.tv_series))
-    Column(
-        Modifier
-            .fillMaxWidth()
-            .background(Color.Gray)
-    ) {
-        TabRow(modifier = Modifier.background(Color.White),
-            selectedTabIndex = selectedTabIndex,
-            indicator = { tabPositions ->
-                TabRowDefaults.PrimaryIndicator(
-                    Modifier.tabIndicatorOffset(tabPositions[selectedTabIndex]),
-                    color = MaterialTheme.colorScheme.primary
-                )
-            }) {
-            tabs.forEachIndexed { index, title ->
-                Tab(selected = selectedTabIndex == index, onClick = {
-                    selectedTabIndex = index
-                    if (index == ACTIVE_MOVIE_TAB) {
-                        navigator.singleTopNavigator(Screen.FavoriteMovie.route)
-                    } else if (index == ACTIVE_TV_SERIES_TAB) {
-                        navigator.singleTopNavigator(Screen.FavoriteTvSeries.route)
-                    }
-
-                }, text = {
-                    Text(
-                        title,
-                        color = if (selectedTabIndex == index) MaterialTheme.colorScheme.primary else Color.Gray
-                    )
-                })
-            }
         }
     }
 }
